Modify: db 변경 사항 반영 room 테이블 - price: int ->  float (다른 통화 고려) - baths: int ->  float  (1.5, 2.5 bath 로 표시되는 경우 고려) - description :  varchar -> text
review table 테이블
- score: int -> float

+++ database/schema/09.create_review.sql
@@ -1,7 +1,7 @@
 CREATE TABLE review(
   id INT NOT NULL AUTO_INCREMENT,
   review text NOT NULL,
-  score INT,
+  score float,
   user_id INT NOT NULL,
   room_id INT NOT NULL,
   reservation_id INT NOT NULL,
